wing support updates
