+++ behringer_mixer/mixers/mixer_type_xr12.py
@@ -1,0 +1,12 @@
+from .mixer_type_xair import MixerTypeXAir
+
+
+class MixerTypeXR12(MixerTypeXAir):
+    """Class for Behringer XR-12 Mixer"""
+
+    mixer_type: str = "XR12"
+    num_channel: int = 12
+    num_bus: int = 2
+    num_dca: int = 4
+    num_fx: int = 4
+    num_head_amp: int = 4
